+++ native/runtime_libc/include/assert.h
@@ -1,0 +1,3 @@
+/* Cide assert.h stub */
+void __cide_assert_fail(void);
+#define assert(expr) if (!(expr)) __cide_assert_fail()