+++ native/runtime_libc/include/errno.h
@@ -1,0 +1,8 @@
+/* Cide errno.h stub */
+extern int errno;
+
+#define EINVAL 1
+#define ERANGE 2
+#define EDOM 3
+#define ENOENT 4
+#define EACCES 5