+++ native/runtime_libc/include/time.h
@@ -1,0 +1,8 @@
+/* Cide time.h stub */
+typedef long long time_t;
+typedef long long clock_t;
+
+#define CLOCKS_PER_SEC 1000000
+
+time_t time(time_t* tloc);
+clock_t clock(void);